TTS/requirements.txt
gTTS
paho-mqtt
